+++ lib/src/widgets/chat_list.dart
@@ -270,7 +270,7 @@ class _ChatListState extends State<ChatList>
         final message = item['message']! as types.Message;
 
         // Compare items to fire only on newly added messages.
-        if (oldMessage != message) {
+        if (oldMessage.id != message.id) {
           // Run only for sent message.
           if (message.author.id == InheritedUser.of(context).user.id) {
             // Delay to give some time for Flutter to calculate new