+++ lib/src/widgets/message/message.dart
@@ -21,6 +21,7 @@ class Message extends StatelessWidget {
   /// Creates a particular message from any message type.
   const Message({
     super.key,
+    this.audioMessageBuilder,
     this.avatarBuilder,
     this.bubbleBuilder,
     this.bubbleRtlAlignment,
@@ -51,7 +52,12 @@ class Message extends StatelessWidget {
     required this.textMessageOptions,
     required this.usePreviewData,
     this.userAgent,
+    this.videoMessageBuilder,
   });
+
+  /// Build an audio message inside predefined bubble.
+  final Widget Function(types.AudioMessage, {required int messageWidth})?
+      audioMessageBuilder;
 
   /// This is to allow custom user avatar builder
   /// By using this we can fetch newest user info based on id
@@ -164,6 +170,10 @@ class Message extends StatelessWidget {
 
   /// See [TextMessage.userAgent].
   final String? userAgent;
+
+  /// Build an audio message inside predefined bubble.
+  final Widget Function(types.VideoMessage, {required int messageWidth})?
+      videoMessageBuilder;
 
   @override
   Widget build(BuildContext context) {
@@ -323,6 +333,11 @@ class Message extends StatelessWidget {
 
   Widget _messageBuilder() {
     switch (message.type) {
+      case types.MessageType.audio:
+        final audioMessage = message as types.AudioMessage;
+        return audioMessageBuilder != null
+            ? audioMessageBuilder!(audioMessage, messageWidth: messageWidth)
+            : const SizedBox();
       case types.MessageType.custom:
         final customMessage = message as types.CustomMessage;
         return customMessageBuilder != null
@@ -361,6 +376,11 @@ class Message extends StatelessWidget {
                 usePreviewData: usePreviewData,
                 userAgent: userAgent,
               );
+      case types.MessageType.video:
+        final videoMessage = message as types.VideoMessage;
+        return videoMessageBuilder != null
+            ? videoMessageBuilder!(videoMessage, messageWidth: messageWidth)
+            : const SizedBox();
       default:
         return const SizedBox();
     }